image/requirements.txt
boto3==1.34.117
fastapi
mangum==0.17.0
uvicorn==0.30.0
langchain==0.2.1
langchain-aws==0.1.6
PyPDF2
pinecone-client
python-dotenv
tavily-python
langgraph
deep_translator